fix(config): tighten pass_up count to uint8, drop dup check
Mirrors the Rust u8 on PassUpConfig (bonus.rs). A count over 255 now fails at Go YAML unmarshal instead of surfacing as a ValidationError, so the rules.go > 255 block is dead code and is removed; the < 1 check stays (0 is a valid uint8 but an invalid pass-up count). TestValidatePassUp_CountExceeds255 is deleted - its scenario is unrepresentable now that the type enforces the bound - replaced by unmarshal-level boundary tests (255 ok; 256/300/-1 rejected).

+++ internal/config/types.go
@@ -600,8 +600,8 @@ type BoardCyclingConfig struct {
 
 // PassUpConfig defines pass-up bonus behavior where initial sales go to the upline.
 type PassUpConfig struct {
-	Count               int  `yaml:"count" json:"count"`
-	IncludesCommissions bool `yaml:"includes_commissions" json:"includes_commissions"`
+	Count               uint8 `yaml:"count" json:"count"`
+	IncludesCommissions bool  `yaml:"includes_commissions" json:"includes_commissions"`
 }
 
 // --- Payout ---
